refactor: user/analytics コンテキストを新設し admin ルートを分離
admin-users を contexts/user/ に、admin-analytics を contexts/analytics/ に移動。
shared に残すのは横断的関心事（admin-auth middleware）と
複数コンテキストをまたぐ集計ビュー（admin-dashboard）のみ。
dep-cruise にコンテキスト分離ルールも追加。
