+++ apps/server/src/app.ts
@@ -1,4 +1,5 @@
 import { Hono } from 'hono';
+import { adminAnalytics } from './contexts/analytics/presentation/routes/admin-analytics.js';
 import { board } from './contexts/board/presentation/routes/board.js';
 import { entries } from './contexts/entry/presentation/routes/entries.js';
 import { adminFermentations } from './contexts/fermentation/presentation/routes/admin-fermentations.js';
@@ -8,10 +9,9 @@ import { questions } from './contexts/question/presentation/routes/questions.js'
 import { adminAuthMiddleware } from './contexts/shared/presentation/middleware/admin-auth.js';
 import { authMiddleware } from './contexts/shared/presentation/middleware/auth.js';
 import { errorHandler } from './contexts/shared/presentation/middleware/error-handler.js';
-import { adminAnalytics } from './contexts/shared/presentation/routes/admin-analytics.js';
 import { adminDashboard } from './contexts/shared/presentation/routes/admin-dashboard.js';
-import { adminUsers } from './contexts/shared/presentation/routes/admin-users.js';
 import { authRoutes } from './contexts/shared/presentation/routes/auth.js';
+import { adminUsers } from './contexts/user/presentation/routes/admin-users.js';
 
 const app = new Hono()
   .onError(errorHandler)
