+++ apps/server/src/contexts/analytics/presentation/routes/admin-analytics.ts
